split out embedding into search
